Change maxCount parameter to int for idiomatic Go channel limits
- NewBoundedChan now takes maxCount as int (not int64) to match Go idioms for channel/slice capacity
- Removed unnecessary math.MaxInt validation
- Updated test panic messages to match lo.Assert format
- Fix TestConcurrentCapReads: add runtime.Gosched() and time.Sleep to ensure goroutines execute
- All tests pass with -race detector enabled

+++ synctools/bounded_chan_test.go
@@ -33,7 +33,7 @@ func (s *boundedChanTestSuite) TestBasicSendReceive() {
 }
 
 func (s *boundedChanTestSuite) TestCountLimitEnforced() {
-	maxCount := int64(3)
+	maxCount := 3
 	out, in, _ := NewBoundedChan(maxCount, 10000, func(i int) int64 { return 1 })
 
 	// Send maxCount + 1 items in a goroutine to avoid deadlock
